Made ServiceResources common.
